refactor(semver): remove internal package, use package-private
Move all implementation classes from internal subpackage to main package
and make them package-private. This provides stronger encapsulation than
the internal package pattern since package-private cannot be bypassed
without reflection.

Package-private is the Java default for a reason - it provides natural
encapsulation boundaries. Classes that don't need to be public are now
only accessible within the package.
