+++ module/semver/src/main/java/com/xenoterracide/gradle/semver/AfterTagHeadStrategy.java
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
 
-package com.xenoterracide.gradle.semver.internal;
+package com.xenoterracide.gradle.semver;
 
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -19,7 +19,7 @@ import org.semver4j.Semver;
  *
  * <p>No build metadata is added on the HEAD branch to keep versions clean.</p>
  */
-public final class AfterTagHeadStrategy implements VersionStrategy {
+final class AfterTagHeadStrategy implements VersionStrategy {
 
   @Override
   public Semver calculate(GitContext ctx) {
