+++ module/semver/src/main/java/com/xenoterracide/gradle/semver/VersionStrategyFactory.java
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
 
-package com.xenoterracide.gradle.semver.internal;
+package com.xenoterracide.gradle.semver;
 
 import org.semver4j.Semver;
 
@@ -15,7 +15,7 @@ import org.semver4j.Semver;
  *   <li>Branch type: HEAD_BRANCH or TOPIC_BRANCH</li>
  * </ul>
  */
-public final class VersionStrategyFactory {
+final class VersionStrategyFactory {
 
   private VersionStrategyFactory() {
     // utility class
