--- module/semver/src/main/java/com/xenoterracide/gradle/semver/internal/package-info.java
@@ -1,11 +1,0 @@
-// SPDX-FileCopyrightText: Copyright © 2024-2026 Caleb Cushing
-//
-// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
-
-/**
- * Internal state machine implementation for semantic version calculation.
- */
-@NullMarked
-package com.xenoterracide.gradle.semver.internal;
-
-import org.jspecify.annotations.NullMarked;
